tests: update unit tests
